refactor: Move tombstones_to_delete_predicates to the predicate crate
