+++ data_types2/src/lib.rs
@@ -11,7 +11,6 @@
 )]
 
 use influxdb_line_protocol::FieldValue;
-use predicate::delete_predicate::parse_delete_predicate;
 use schema::{builder::SchemaBuilder, sort::SortKey, InfluxColumnType, InfluxFieldType, Schema};
 use std::{
     collections::BTreeMap,
@@ -697,27 +696,6 @@ pub struct Tombstone {
     pub max_time: Timestamp,
     /// the full delete predicate
     pub serialized_predicate: String,
-}
-
-/// Convert tombstones to delete predicates
-pub fn tombstones_to_delete_predicates(tombstones: &[Tombstone]) -> Vec<Arc<DeletePredicate>> {
-    tombstones_to_delete_predicates_iter(tombstones).collect()
-}
-
-/// Return Iterator of delete predicates
-pub fn tombstones_to_delete_predicates_iter(
-    tombstones: &[Tombstone],
-) -> impl Iterator<Item = Arc<DeletePredicate>> + '_ {
-    tombstones.iter().map(|tombstone| {
-        Arc::new(
-            parse_delete_predicate(
-                &tombstone.min_time.get().to_string(),
-                &tombstone.max_time.get().to_string(),
-                &tombstone.serialized_predicate,
-            )
-            .expect("Error building delete predicate"),
-        )
-    })
 }
 
 /// Data for a parquet file reference that has been inserted in the catalog.
