+++ predicate/Cargo.toml
@@ -7,6 +7,7 @@ edition = "2021"
 arrow = { version = "13", features = ["prettyprint"] }
 chrono = { version = "0.4", default-features = false }
 data_types = { path = "../data_types" }
+data_types2 = { path = "../data_types2" }
 datafusion = { path = "../datafusion" }
 datafusion_util = { path = "../datafusion_util" }
 itertools = "0.10"
